# Use custom d-ary heap implementation