# Context
## Benchmark highlights
CH Map Matching +16% ops/sec
MLD Map matching +30% ops/sec
CH Routing +3% ops/sec
MLD Routing +18% ops/sec
CH Table +7% ops/sec
MLD Table +16% ops/sec

osrm_contract run time on berlin-latest.osm.pbf -24%
osrm_customize run time on berlin-latest.osm.pbf -25%


## Description
It turns out that with mutability enabled `boost::heap::d_ary_heap` becomes extremely ineffective. I wrote such benchmark which compares `boost::heap::mutable_<true>` vs `boost::heap::mutable_<false>`:
```
    const int NUM_ELEMENTS = 10000000;
    std::vector<uint16_t> data;
    data.reserve(NUM_ELEMENTS);

    std::mt19937 gen(42);
    std::uniform_int_distribution<> dis(1, 1000000);

    for (int i = 0; i < NUM_ELEMENTS; ++i) {
        data.push_back(dis(gen));
    }

    std::vector<int> resultBoostDAryHeapMutable;
    resultBoostDAryHeapMutable.reserve(NUM_ELEMENTS);
    auto start = std::chrono::high_resolution_clock::now();
    boost::heap::d_ary_heap<uint16_t, boost::heap::arity<4>, boost::heap::mutable_<true>, boost::heap::compare<std::greater<uint16_t>>> boostDAryHeapMutable;
    for (const auto& val : data) {
        boostDAryHeapMutable.push(val);
    }
    while (!boostDAryHeapMutable.empty()) {
        resultBoostDAryHeapMutable.push_back(boostDAryHeapMutable.top());
        boostDAryHeapMutable.pop();
    }
    auto end = std::chrono::high_resolution_clock::now();
    auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
    std::cout << "Boost d_ary_heap mutable took " << duration << " ms" << std::endl;

    std::vector<int> resultBoostDAryHeap;
    resultBoostDAryHeap.reserve(NUM_ELEMENTS);
    start = std::chrono::high_resolution_clock::now();
    boost::heap::d_ary_heap<uint16_t, boost::heap::arity<4>, boost::heap::compare<std::greater<uint16_t>>> boostDAryHeap;
    for (const auto& val : data) {
        boostDAryHeap.push(val);
    }
    while (!boostDAryHeap.empty()) {
        resultBoostDAryHeap.push_back(boostDAryHeap.top());
        boostDAryHeap.pop();
    }
    end = std::chrono::high_resolution_clock::now();
    duration = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
    std::cout << "Boost d_ary_heap took " << duration << " ms" << std::endl;
}
```

And it turns out that mutable implementation(i.e. where we can decrease priorities) is 10-12x(!) slower than non-mutable one:
```
Boost d_ary_heap mutable took 11142 ms
Boost d_ary_heap took 901 ms
```

Implementation of boost::heap::d_ary_heap is quite difficult to read, but as I understood in order to achieve mutability it [uses](https://www.boost.org/doc/libs/1_84_0/boost/heap/detail/mutable_heap.hpp) `std::list` under the hood what obviously makes it quite slow(at least due to additional allocations and non cache-friendliness). 
So I decided to make our own implementation of 4-ary heap which handles mutability in a different way: instead of having stable iterators we use indexes in heap's underlying vector with a bit of additional logic to update stored indexes when elements are reordered within the heap.





<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 11076<br/>plain u32: 10972.9<br/>aliased double: 15129.3<br/>plain double: 15113.2 | aliased u32: 11009.8<br/>plain u32: 10996.4<br/>aliased double: 15039.1<br/>plain double: 15025.3 |
| e2e_match_ch | Ops: 25.55 ± 0.01 ops/s. Best: 25.57 ops/s<br/>Total: 5126.51ms ± 2.43ms. Best: 5122.78ms<br/>Min time: 3.05ms ± 0.05ms<br/>Mean time: 39.13ms ± 0.02ms<br/>Median time: 27.37ms ± 0.08ms<br/>95th percentile: 132.54ms ± 0.35ms<br/>99th percentile: 158.85ms ± 0.46ms<br/>Max time: 179.68ms ± 0.75ms | Ops: 30.69 ± 0.03 ops/s. Best: 30.75 ops/s<br/>Total: 4268.10ms ± 5.06ms. Best: 4259.72ms<br/>Min time: 2.93ms ± 0.05ms<br/>Mean time: 32.58ms ± 0.04ms<br/>Median time: 22.89ms ± 0.14ms<br/>95th percentile: 106.88ms ± 0.12ms<br/>99th percentile: 128.44ms ± 0.57ms<br/>Max time: 141.70ms ± 0.89ms |
| e2e_match_mld | Ops: 43.57 ± 0.02 ops/s. Best: 43.60 ops/s<br/>Total: 3006.32ms ± 1.34ms. Best: 3004.43ms<br/>Min time: 2.58ms ± 0.05ms<br/>Mean time: 22.95ms ± 0.01ms<br/>Median time: 12.06ms ± 0.12ms<br/>95th percentile: 75.20ms ± 0.12ms<br/>99th percentile: 86.91ms ± 0.14ms<br/>Max time: 100.71ms ± 0.30ms | Ops: 59.95 ± 0.05 ops/s. Best: 60.03 ops/s<br/>Total: 2184.99ms ± 1.71ms. Best: 2182.09ms<br/>Min time: 2.24ms ± 0.02ms<br/>Mean time: 16.68ms ± 0.01ms<br/>Median time: 9.72ms ± 0.04ms<br/>95th percentile: 52.05ms ± 0.10ms<br/>99th percentile: 60.11ms ± 0.16ms<br/>Max time: 69.53ms ± 0.39ms |
| e2e_nearest_ch | Ops: 635.40 ± 2.55 ops/s. Best: 640.86 ops/s<br/>Total: 1573.68ms ± 6.29ms. Best: 1560.39ms<br/>Min time: 1.28ms ± 0.01ms<br/>Mean time: 1.57ms ± 0.01ms<br/>Median time: 1.54ms ± 0.00ms<br/>95th percentile: 1.96ms ± 0.01ms<br/>99th percentile: 2.05ms ± 0.01ms<br/>Max time: 9.38ms ± 7.45ms | Ops: 636.59 ± 3.36 ops/s. Best: 642.66 ops/s<br/>Total: 1570.80ms ± 8.30ms. Best: 1556.04ms<br/>Min time: 1.28ms ± 0.01ms<br/>Mean time: 1.57ms ± 0.01ms<br/>Median time: 1.53ms ± 0.01ms<br/>95th percentile: 1.96ms ± 0.01ms<br/>99th percentile: 2.04ms ± 0.01ms<br/>Max time: 9.32ms ± 7.40ms |
| e2e_nearest_mld | Ops: 635.02 ± 4.27 ops/s. Best: 640.52 ops/s<br/>Total: 1574.46ms ± 10.06ms. Best: 1561.23ms<br/>Min time: 1.29ms ± 0.01ms<br/>Mean time: 1.57ms ± 0.01ms<br/>Median time: 1.54ms ± 0.01ms<br/>95th percentile: 1.97ms ± 0.01ms<br/>99th percentile: 2.05ms ± 0.01ms<br/>Max time: 9.39ms ± 7.43ms | Ops: 636.33 ± 2.53 ops/s. Best: 639.88 ops/s<br/>Total: 1571.45ms ± 6.62ms. Best: 1562.79ms<br/>Min time: 1.28ms ± 0.01ms<br/>Mean time: 1.57ms ± 0.01ms<br/>Median time: 1.54ms ± 0.00ms<br/>95th percentile: 1.96ms ± 0.00ms<br/>99th percentile: 2.05ms ± 0.01ms<br/>Max time: 9.28ms ± 7.37ms |
| e2e_route_ch | Ops: 217.14 ± 0.62 ops/s. Best: 218.12 ops/s<br/>Total: 4605.59ms ± 13.80ms. Best: 4584.62ms<br/>Min time: 1.83ms ± 0.03ms<br/>Mean time: 4.61ms ± 0.01ms<br/>Median time: 4.70ms ± 0.01ms<br/>95th percentile: 6.06ms ± 0.03ms<br/>99th percentile: 6.60ms ± 0.04ms<br/>Max time: 13.34ms ± 6.42ms | Ops: 218.93 ± 0.47 ops/s. Best: 219.97 ops/s<br/>Total: 4567.70ms ± 10.21ms. Best: 4546.15ms<br/>Min time: 1.84ms ± 0.05ms<br/>Mean time: 4.57ms ± 0.01ms<br/>Median time: 4.66ms ± 0.01ms<br/>95th percentile: 6.01ms ± 0.03ms<br/>99th percentile: 6.55ms ± 0.03ms<br/>Max time: 13.24ms ± 6.37ms |
| e2e_route_mld | Ops: 178.18 ± 0.27 ops/s. Best: 178.71 ops/s<br/>Total: 5612.39ms ± 8.61ms. Best: 5595.71ms<br/>Min time: 1.78ms ± 0.04ms<br/>Mean time: 5.61ms ± 0.01ms<br/>Median time: 5.73ms ± 0.02ms<br/>95th percentile: 7.61ms ± 0.03ms<br/>99th percentile: 8.15ms ± 0.04ms<br/>Max time: 14.77ms ± 5.89ms | Ops: 186.43 ± 0.32 ops/s. Best: 187.08 ops/s<br/>Total: 5363.86ms ± 9.38ms. Best: 5345.20ms<br/>Min time: 1.78ms ± 0.03ms<br/>Mean time: 5.36ms ± 0.01ms<br/>Median time: 5.47ms ± 0.02ms<br/>95th percentile: 7.25ms ± 0.01ms<br/>99th percentile: 7.82ms ± 0.06ms<br/>Max time: 14.42ms ± 5.98ms |
| e2e_table_ch | Ops: 218.31 ± 0.48 ops/s. Best: 219.32 ops/s<br/>Total: 4580.74ms ± 10.97ms. Best: 4559.55ms<br/>Min time: 2.46ms ± 0.04ms<br/>Mean time: 4.58ms ± 0.01ms<br/>Median time: 4.58ms ± 0.01ms<br/>95th percentile: 6.23ms ± 0.02ms<br/>99th percentile: 6.58ms ± 0.04ms<br/>Max time: 13.70ms ± 7.04ms | Ops: 226.94 ± 0.38 ops/s. Best: 227.56 ops/s<br/>Total: 4406.36ms ± 7.58ms. Best: 4394.42ms<br/>Min time: 2.40ms ± 0.03ms<br/>Mean time: 4.41ms ± 0.01ms<br/>Median time: 4.40ms ± 0.02ms<br/>95th percentile: 5.96ms ± 0.02ms<br/>99th percentile: 6.29ms ± 0.03ms<br/>Max time: 13.45ms ± 7.07ms |
| e2e_table_mld | Ops: 69.13 ± 0.03 ops/s. Best: 69.16 ops/s<br/>Total: 14466.36ms ± 6.86ms. Best: 14460.18ms<br/>Min time: 6.01ms ± 0.04ms<br/>Mean time: 14.47ms ± 0.01ms<br/>Median time: 14.39ms ± 0.01ms<br/>95th percentile: 22.07ms ± 0.04ms<br/>99th percentile: 23.22ms ± 0.08ms<br/>Max time: 29.71ms ± 5.67ms | Ops: 78.73 ± 0.02 ops/s. Best: 78.76 ops/s<br/>Total: 12701.45ms ± 4.01ms. Best: 12696.81ms<br/>Min time: 5.38ms ± 0.02ms<br/>Mean time: 12.70ms ± 0.00ms<br/>Median time: 12.65ms ± 0.01ms<br/>95th percentile: 19.31ms ± 0.06ms<br/>99th percentile: 20.29ms ± 0.06ms<br/>Max time: 26.86ms ± 5.89ms |
| e2e_trip_ch | Ops: 62.95 ± 0.03 ops/s. Best: 63.00 ops/s<br/>Total: 15885.05ms ± 8.02ms. Best: 15874.26ms<br/>Min time: 2.43ms ± 0.22ms<br/>Mean time: 15.88ms ± 0.01ms<br/>Median time: 15.10ms ± 0.03ms<br/>95th percentile: 28.09ms ± 0.03ms<br/>99th percentile: 30.17ms ± 0.09ms<br/>Max time: 32.49ms ± 1.45ms | Ops: 64.23 ± 0.04 ops/s. Best: 64.29 ops/s<br/>Total: 15568.75ms ± 9.29ms. Best: 15553.72ms<br/>Min time: 2.38ms ± 0.17ms<br/>Mean time: 15.57ms ± 0.01ms<br/>Median time: 14.78ms ± 0.03ms<br/>95th percentile: 27.65ms ± 0.05ms<br/>99th percentile: 29.60ms ± 0.09ms<br/>Max time: 32.03ms ± 1.63ms |
| e2e_trip_mld | Ops: 36.66 ± 0.01 ops/s. Best: 36.68 ops/s<br/>Total: 27281.21ms ± 8.68ms. Best: 27263.72ms<br/>Min time: 2.42ms ± 0.17ms<br/>Mean time: 27.28ms ± 0.01ms<br/>Median time: 26.38ms ± 0.05ms<br/>95th percentile: 44.34ms ± 0.05ms<br/>99th percentile: 47.07ms ± 0.12ms<br/>Max time: 49.41ms ± 0.11ms | Ops: 40.05 ± 0.02 ops/s. Best: 40.08 ops/s<br/>Total: 24966.90ms ± 14.12ms. Best: 24951.03ms<br/>Min time: 2.45ms ± 0.19ms<br/>Mean time: 24.97ms ± 0.01ms<br/>Median time: 24.13ms ± 0.07ms<br/>95th percentile: 41.08ms ± 0.04ms<br/>99th percentile: 43.63ms ± 0.11ms<br/>Max time: 45.98ms ± 0.42ms |
| json-render | String: 8.89322ms<br/>Stringstream: 14.6254ms<br/>Vector: 9.48771ms | String: 9.15423ms<br/>Stringstream: 14.6142ms<br/>Vector: 9.53894ms |
| match_ch | Default radius: <br/>7.06186ms/req at 82 coordinate<br/>0.0861202ms/coordinate<br/>Radius 10m: <br/>25.0142ms/req at 82 coordinate<br/>0.305052ms/coordinate | Default radius: <br/>6.28653ms/req at 82 coordinate<br/>0.076665ms/coordinate<br/>Radius 10m: <br/>22.18ms/req at 82 coordinate<br/>0.270488ms/coordinate |
| match_mld | Default radius: <br/>4.35379ms/req at 82 coordinate<br/>0.053095ms/coordinate<br/>Radius 10m: <br/>16.2432ms/req at 82 coordinate<br/>0.198088ms/coordinate | Default radius: <br/>3.11549ms/req at 82 coordinate<br/>0.0379938ms/coordinate<br/>Radius 10m: <br/>10.7051ms/req at 82 coordinate<br/>0.13055ms/coordinate |
| node_match_ch | Ops: 159.3 ± 0.5 ops/s. Best: 160.0 ops/s | Ops: 186.1 ± 1.3 ops/s. Best: 187.3 ops/s |
| node_match_mld | Ops: 219.5 ± 2.2 ops/s. Best: 221.6 ops/s | Ops: 286.7 ± 1.7 ops/s. Best: 290.0 ops/s |
| node_nearest_ch | Ops: 9989.3 ± 821.8 ops/s. Best: 11065.2 ops/s | Ops: 9393.8 ± 443.1 ops/s. Best: 10209.7 ops/s |
| node_nearest_mld | Ops: 9286.5 ± 369.8 ops/s. Best: 9970.7 ops/s | Ops: 10275.8 ± 858.9 ops/s. Best: 12036.0 ops/s |
| node_route_ch | Ops: 1002.6 ± 25.0 ops/s. Best: 1039.0 ops/s | Ops: 1031.5 ± 23.5 ops/s. Best: 1057.3 ops/s |
| node_route_mld | Ops: 506.0 ± 6.3 ops/s. Best: 515.4 ops/s | Ops: 585.5 ± 3.6 ops/s. Best: 590.5 ops/s |
| node_table_ch | Ops: 186.4 ± 1.9 ops/s. Best: 188.9 ops/s | Ops: 197.6 ± 1.9 ops/s. Best: 200.9 ops/s |
| node_table_mld | Ops: 37.8 ± 0.1 ops/s. Best: 37.9 ops/s | Ops: 44.0 ± 0.1 ops/s. Best: 44.1 ops/s |
| node_trip_ch | Ops: 185.9 ± 1.0 ops/s. Best: 187.0 ops/s | Ops: 193.7 ± 0.4 ops/s. Best: 194.6 ops/s |
| node_trip_mld | Ops: 61.4 ± 0.2 ops/s. Best: 61.6 ops/s | Ops: 71.8 ± 0.3 ops/s. Best: 72.1 ops/s |
| osrm_contract | Time: 184.30s Peak RAM: 195.01MB | Time: 141.54s Peak RAM: 195.00MB |
| osrm_customize | Time: 2.53s Peak RAM: 112.45MB | Time: 1.92s Peak RAM: 112.45MB |
| osrm_extract | Time: 24.33s Peak RAM: 398.68MB | Time: 24.26s Peak RAM: 399.01MB |
| osrm_partition | Time: 5.98s Peak RAM: 119.62MB | Time: 5.95s Peak RAM: 119.43MB |
| packedvector | random write:<br/>std::vector 184015 ms<br/>util::packed_vector 373595 ms<br/>slowdown: 2.03025<br/>random read:<br/>std::vector 100548 ms<br/>util::packed_vector 190832 ms<br/>slowdown: 1.89791 | random write:<br/>std::vector 184979 ms<br/>util::packed_vector 373751 ms<br/>slowdown: 2.02051<br/>random read:<br/>std::vector 100553 ms<br/>util::packed_vector 190728 ms<br/>slowdown: 1.89679 |
| random_match_ch | 500 matches, default radius<br/>ops: 120.21 ± 0.19 ops/s. best: 120.49ops/s.<br/>total: 474.18 ± 0.76ms. best: 473.05ms.<br/>avg: 8.32 ± 0.01ms<br/>min: 0.23 ± 0.01ms<br/>max: 41.83 ± 0.17ms<br/>p99: 41.83 ± 0.17ms<br/><br/>500 matches, radius=10<br/>ops: 34.48 ± 0.05 ops/s. best: 34.53ops/s.<br/>total: 1856.08 ± 2.44ms. best: 1853.21ms.<br/>avg: 29.00 ± 0.04ms<br/>min: 0.22 ± 0.00ms<br/>max: 412.12 ± 1.39ms<br/>p99: 412.12 ± 1.39ms<br/><br/>500 matches, radius=20<br/>ops: 8.11 ± 0.01 ops/s. best: 8.12ops/s.<br/>total: 8019.57 ± 8.18ms. best: 8009.41ms.<br/>avg: 123.38 ± 0.13ms<br/>min: 0.49 ± 0.00ms<br/>max: 2128.87 ± 3.56ms<br/>p99: 2128.87 ± 3.56ms<br/><br/>Peak RAM: 55.000MB | 500 matches, default radius<br/>ops: 146.70 ± 0.27 ops/s. best: 147.08ops/s.<br/>total: 388.54 ± 0.73ms. best: 387.54ms.<br/>avg: 6.82 ± 0.01ms<br/>min: 0.22 ± 0.00ms<br/>max: 34.01 ± 0.07ms<br/>p99: 34.01 ± 0.07ms<br/><br/>500 matches, radius=10<br/>ops: 42.64 ± 0.04 ops/s. best: 42.68ops/s.<br/>total: 1500.93 ± 1.58ms. best: 1499.46ms.<br/>avg: 23.45 ± 0.02ms<br/>min: 0.22 ± 0.00ms<br/>max: 332.39 ± 0.72ms<br/>p99: 332.39 ± 0.72ms<br/><br/>500 matches, radius=20<br/>ops: 10.10 ± 0.01 ops/s. best: 10.12ops/s.<br/>total: 6436.36 ± 8.71ms. best: 6423.04ms.<br/>avg: 99.02 ± 0.13ms<br/>min: 0.43 ± 0.00ms<br/>max: 1691.74 ± 4.69ms<br/>p99: 1691.74 ± 4.69ms<br/><br/>Peak RAM: 55.000MB |
| random_match_mld | 500 matches, default radius<br/>ops: 205.04 ± 0.49 ops/s. best: 205.56ops/s.<br/>total: 278.00 ± 0.67ms. best: 277.29ms.<br/>avg: 4.88 ± 0.01ms<br/>min: 0.21 ± 0.00ms<br/>max: 27.09 ± 0.05ms<br/>p99: 27.09 ± 0.05ms<br/><br/>500 matches, radius=10<br/>ops: 73.10 ± 0.06 ops/s. best: 73.17ops/s.<br/>total: 875.56 ± 0.73ms. best: 874.64ms.<br/>avg: 13.68 ± 0.01ms<br/>min: 0.21 ± 0.00ms<br/>max: 161.38 ± 0.32ms<br/>p99: 161.38 ± 0.32ms<br/><br/>500 matches, radius=20<br/>ops: 15.37 ± 0.01 ops/s. best: 15.38ops/s.<br/>total: 4229.16 ± 2.25ms. best: 4225.76ms.<br/>avg: 65.06 ± 0.03ms<br/>min: 0.29 ± 0.00ms<br/>max: 843.11 ± 1.06ms<br/>p99: 843.11 ± 1.06ms<br/><br/>Peak RAM: 51.500MB | 500 matches, default radius<br/>ops: 305.47 ± 1.12 ops/s. best: 306.62ops/s.<br/>total: 186.60 ± 0.69ms. best: 185.90ms.<br/>avg: 3.27 ± 0.01ms<br/>min: 0.18 ± 0.00ms<br/>max: 18.14 ± 0.05ms<br/>p99: 18.14 ± 0.05ms<br/><br/>500 matches, radius=10<br/>ops: 112.95 ± 0.08 ops/s. best: 113.05ops/s.<br/>total: 566.63 ± 0.38ms. best: 566.14ms.<br/>avg: 8.85 ± 0.01ms<br/>min: 0.20 ± 0.00ms<br/>max: 100.66 ± 0.10ms<br/>p99: 100.66 ± 0.10ms<br/><br/>500 matches, radius=20<br/>ops: 23.63 ± 0.02 ops/s. best: 23.66ops/s.<br/>total: 2750.81 ± 1.83ms. best: 2747.78ms.<br/>avg: 42.32 ± 0.03ms<br/>min: 0.26 ± 0.00ms<br/>max: 536.31 ± 0.44ms<br/>p99: 536.31 ± 0.44ms<br/><br/>Peak RAM: 51.500MB |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>ops: 21344.30 ± 41.26 ops/s. best: 21377.87ops/s.<br/>total: 468.51 ± 0.91ms. best: 467.77ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.16 ± 0.03ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 15599.62 ± 3.63 ops/s. best: 15604.45ops/s.<br/>total: 641.04 ± 0.15ms. best: 640.84ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 11959.99 ± 3.89 ops/s. best: 11965.45ops/s.<br/>total: 836.12 ± 0.27ms. best: 835.74ms.<br/>avg: 0.08 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 0.19 ± 0.00ms<br/>p99: 0.15 ± 0.00ms<br/><br/>Peak RAM: 34.500MB | 10000 nearest, number_of_results=1<br/>ops: 21446.43 ± 39.13 ops/s. best: 21477.18ops/s.<br/>total: 466.28 ± 0.85ms. best: 465.61ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.02ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 15616.52 ± 12.28 ops/s. best: 15628.15ops/s.<br/>total: 640.35 ± 0.50ms. best: 639.87ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.16 ± 0.02ms<br/>p99: 0.13 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 11960.74 ± 4.46 ops/s. best: 11965.85ops/s.<br/>total: 836.07 ± 0.31ms. best: 835.71ms.<br/>avg: 0.08 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 0.19 ± 0.00ms<br/>p99: 0.15 ± 0.00ms<br/><br/>Peak RAM: 34.500MB |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>ops: 21347.69 ± 37.14 ops/s. best: 21374.15ops/s.<br/>total: 468.44 ± 0.82ms. best: 467.85ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.02ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 15600.41 ± 6.40 ops/s. best: 15606.31ops/s.<br/>total: 641.01 ± 0.26ms. best: 640.77ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.13 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 11957.44 ± 6.00 ops/s. best: 11963.98ops/s.<br/>total: 836.30 ± 0.43ms. best: 835.84ms.<br/>avg: 0.08 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 0.19 ± 0.00ms<br/>p99: 0.15 ± 0.00ms<br/><br/>Peak RAM: 34.500MB | 10000 nearest, number_of_results=1<br/>ops: 21445.62 ± 32.89 ops/s. best: 21472.30ops/s.<br/>total: 466.30 ± 0.72ms. best: 465.72ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.02ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 15618.99 ± 6.24 ops/s. best: 15631.75ops/s.<br/>total: 640.25 ± 0.26ms. best: 639.72ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 11947.05 ± 5.39 ops/s. best: 11952.48ops/s.<br/>total: 837.03 ± 0.39ms. best: 836.65ms.<br/>avg: 0.08 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 0.19 ± 0.00ms<br/>p99: 0.15 ± 0.00ms<br/><br/>Peak RAM: 34.500MB |
…